requirements.txt
anyio==4.8.0
apptools==5.3.0
async-lru==2.0.4
asyncio-atexit==1.0.1
certifi==2025.1.31
charset-normalizer==3.4.1
configobj==5.0.9
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.13
Deprecated==1.2.18
easydict==1.9
envisage==7.0.3
exceptiongroup==1.2.2
fonttools==4.56.0
ftfy==6.1.3
googleapis-common-protos==1.69.0
grad-cam==1.4.3
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.5.2
joblib==1.4.2
kiwisolver==1.4.8
matplotlib==3.8.4
mayavi==4.8.2
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
opencv-python==4.6.0.66
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-httpx==0.51b0
opentelemetry-instrumentation-requests==0.51b0
opentelemetry-instrumentation-threading==0.51b0
opentelemetry-instrumentation-urllib==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
packaging==24.2
Pillow==9.1.1
plotly==6.0.0
protobuf==5.29.3
psutil==5.9.8
pyface==8.0.0
Pygments==2.19.1
pyparsing==3.2.1
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dateutil==2.9.0.post0
PyWavelets==1.6.0
PyYAML==6.0
regex==2023.12.25
requests==2.32.3
scipy==1.13.1
shapely==2.0.7
six==1.17.0
sklearn==0.0
sniffio==1.3.1
synapseclient==4.7.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tqdm==4.67.1
traits==7.0.2
traitsui==8.0.0
ttach==0.0.3
typing_extensions==4.12.2
urllib3==1.26.20
vtk==9.4.1
wcwidth==0.2.13
wrapt==1.17.2
zipp==3.21.0
scikit-image==0.19.3
scikit-learn==1.1.2
numpy==1.22.4
torch==1.13.0
torchvision==0.14.0